+++ Assets/Simulation/HexMap/IHexGrid.cs
@@ -43,7 +43,7 @@ namespace Assets.Simulation.HexMap {
 
         List<IHexCell> GetShortestPathBetween(IHexCell start, IHexCell end);
 
-        List<IHexCell> GetShortestPathBetween(IHexCell start, IHexCell end, Func<IHexCell, IHexCell, int> costFunction);
+        List<IHexCell> GetShortestPathBetween(IHexCell start, IHexCell end, Func<IHexCell, IHexCell, float> costFunction);
 
         List<IHexCell> GetNeighbors(IHexCell center);
